refactor: react query 모아두는 폴더 이름 변경
