+++ apps/web/src/app/[locale]/shop/SellListSection/EditModal.tsx
@@ -2,11 +2,10 @@
 
 import React, { useState } from 'react';
 import { css } from '_panda/css';
+import { auctionQueries, useChangeProductPrice, useDeleteProduct } from '@gitanimals/react-query';
 import { Button, Modal } from '@gitanimals/ui-panda';
 import { useQueryClient } from '@tanstack/react-query';
 import { toast } from 'sonner';
-
-import { auctionQueries, useChangeProductPrice, useDeleteProduct } from '@gitanimals/react-query';
 
 function EditModal({ isOpen, onClose, productId }: { isOpen: boolean; onClose: () => void; productId?: string }) {
   const queryClient = useQueryClient();
