+++ packages/lib/react-query/src/auction/index.ts
@@ -1,2 +1,2 @@
 export * from './queries';
-export * from './useProduct';
+export * from './product';
