+++ packages/lib/react-query/src/auction/product.ts
