+++ packages/lib/react-query/src/shop/index.ts
@@ -1,2 +1,2 @@
 export * from './queries';
-export * from './useBackground';
+export * from './background';
